Add turrets
Add death ray & gun turrets

Fix out-of-level-bounds handling in terrain_line function

Sort particle array after adding new particles.
If the array is not sorted, iter_slice will not work
correctly and might not render all (or any) particles.
This is normally imperceptible as the array gets sorted on
the next tick, but laser particles only exists for a couple
of ticks.